modified to pick <peak estimates> on scaled data
The large discrepancies between raw values can influence the amplitude and slope thresholds being used. To minimize this effect and make the algorithm more robust, the data are scaled so that they fall between 0 and 1. This scaled version is used to generate the first peak estimates and then these are used with the raw data to refine the peaks on the raw (merged) data.