changed the scaling function
The original scaling function was taking the minimum value that may show up as a negative value. This is impossible and is an artefact of the data and the background removal algorithm. The scaling value now uses the minimum value of (absolute values) as the minimum. This means that we get away of the 'DC' shift that was observed without this modification.