generate PCA tables
Peak statistics was modified to generate tables that are suitable for PCA analysis. The user selects the number of bins to subdivide the data and sets the treshold (i.e. number of observations in a bin required to be considered significant). The user can also set the imputation value which is set to -999 for now.