Threshold is set to 60% of readings
Modifications to merge_peaks so that the number of readings is returned such that we can compute the threshold value (i.e. the number of times a peak should be seen to be kept for analysis). This is set to 60% of the number of readings and is rounded to the lowest value by using int()